refactor/RLT-56: create separate validation to phone input
